requirements.txt
django_localflavor==3.1
Django==3.2.6
django_crispy_forms==1.12.0
